Moved markdown parsing logic to a separate util file
